--- MetalGear/src/Mision.java
@@ -1,12 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-
-/**
- *
- * @author mauro
- */
-public abstract class Mision {
-  public abstract void Iniciar();
-}
