+++ MetalGear/src/metalgear/Celda.java
@@ -1,0 +1,6 @@
+
+package metalgear;
+
+public class Celda {
+    
+}
