+++ MetalGear/src/metalgear/Enemigo.java
@@ -1,0 +1,6 @@
+
+package metalgear;
+
+public interface Enemigo {
+    
+}
