+++ MetalGear/src/metalgear/Guardia.java
@@ -1,0 +1,6 @@
+
+package metalgear;
+
+public class Guardia extends Personaje /*implements Enemigo*/{
+    
+}
