+++ MetalGear/src/metalgear/Juego.java
@@ -1,0 +1,11 @@
+package metalgear;
+
+public class Juego {
+
+    
+  
+    public static void main(String[] args) {
+        
+    }
+    
+}
