+++ MetalGear/src/metalgear/Mapa.java
@@ -1,0 +1,6 @@
+package metalgear;
+
+
+public class Mapa {
+    
+}
