+++ MetalGear/src/metalgear/MetalGear.java
@@ -1,0 +1,5 @@
+package metalgear;
+
+public class MetalGear extends Personaje /*implements Enemigo*/ {
+
+}
