+++ MetalGear/src/metalgear/Mision.java
@@ -1,0 +1,7 @@
+package metalgear;
+
+
+public abstract class Mision {
+    
+  public abstract void Iniciar();
+}
