+++ MetalGear/src/metalgear/MisionFinal.java
@@ -1,0 +1,11 @@
+
+package metalgear;
+
+public class MisionFinal extends Mision{
+
+    @Override
+    public void Iniciar() {
+    
+    }
+    
+}
