+++ MetalGear/src/metalgear/MisionIntermedia.java
@@ -1,0 +1,10 @@
+package metalgear;
+
+public class MisionIntermedia extends Mision {
+
+    @Override
+    public void Iniciar() {
+        
+    }
+
+}
