+++ MetalGear/src/metalgear/Objeto.java
@@ -1,0 +1,6 @@
+
+package metalgear;
+
+public class Objeto {
+    
+}
