+++ MetalGear/src/metalgear/Personaje.java
@@ -1,0 +1,5 @@
+package metalgear;
+
+public abstract class Personaje {
+    
+}
