+++ MetalGear/src/metalgear/Posicion.java
@@ -1,0 +1,6 @@
+
+package metalgear;
+
+public class Posicion {
+    
+}
