+++ MetalGear/src/metalgear/Snake.java
@@ -1,0 +1,6 @@
+
+package metalgear;
+
+public class Snake extends Personaje{
+    
+}
